1.21.4

+++ src/main/java/me/yuugao/meyuugaosradio/Constants.java
@@ -28,13 +28,13 @@ public class Constants {
     public static final String BLOCK_DISMANTLE_SOUND_ID = "block_dismantle";
 
     public static final long RADIO_ENERGY_CAPACITY = 100_000L;
-    public static final int RADIO_ENERGY_USAGE = 8;
+    public static final long RADIO_ENERGY_USAGE = 8L;
 
     public static final long SPEAKER_ENERGY_CAPACITY = 200_000L;
-    public static final int SPEAKER_ENERGY_USAGE = 16;
+    public static final long SPEAKER_ENERGY_USAGE = 16L;
 
     public static final long REMOTE_CONTROLLER_ENERGY_CAPACITY = 10_000L;
-    public static final int REMOTE_CONTROLLER_ENERGY_USAGE = 8;
+    public static final long REMOTE_CONTROLLER_ENERGY_USAGE = 8L;
 
     public static final float RADIO_VOLUME_MULTIPLIER = 2f;
     public static final float SPEAKER_VOLUME_MULTIPLIER = 4f;
